aggregated migrations for the OpenProject 18 squashing
