+++ db/migrate/migration_utils/squashed_migration.rb
@@ -31,7 +31,7 @@
 require Rails.root.join("db/migrate/migration_utils/migration_squasher").to_s
 
 class SquashedMigration < ActiveRecord::Migration[8.1]
-  class_attribute :minimum_version, default: "16"
+  class_attribute :minimum_version, default: "17"
 
   def self.squashed_migrations(*migrations)
     if migrations.any?
