+++ modules/documents/db/migrate/1012017_aggregated_documents_migrations.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# -- copyright
+#-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
 #
@@ -26,28 +26,17 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-# ++
+#++
 
 require Rails.root.join("db/migrate/migration_utils/squashed_migration").to_s
-require_relative "tables/budgets"
-require_relative "tables/budget_journals"
-require_relative "tables/labor_budget_items"
-require_relative "tables/material_budget_items"
+require_relative "tables/documents"
+require_relative "tables/document_journals"
 
-# This migration aggregates the migrations detailed in MIGRATION_FILES
-class AggregatedBudgetMigrations < SquashedMigration
+class AggregatedDocumentsMigrations < SquashedMigration
   squashed_migrations *%w[
-    20200807083950_keep_enabled_module
-    20200810152654_rename_cost_object_to_budget
+    1012015_aggregated_documents_migrations
   ]
 
-  tables Tables::Budgets,
-         Tables::BudgetJournals,
-         Tables::LaborBudgetItems,
-         Tables::MaterialBudgetItems
-
-  modifications do
-    add_column :work_packages, :budget_id, :integer
-    add_column :work_package_journals, :budget_id, :integer, null: true
-  end
+  tables Tables::Documents,
+         Tables::DocumentJournals
 end
