+++ modules/overviews/db/migrate/1019017_aggregated_overviews_migrations.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-#-- copyright
+# -- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
 #
@@ -26,28 +26,12 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-#++
+# ++
 
 require Rails.root.join("db/migrate/migration_utils/squashed_migration").to_s
-require_relative "tables/cost_entries"
-require_relative "tables/cost_types"
-require_relative "tables/rates"
-require_relative "tables/time_entries"
-require_relative "tables/time_entry_activities_projects"
-require_relative "tables/time_entry_journals"
 
-class AggregatedCostsMigrations < SquashedMigration
+class AggregatedOverviewsMigrations < SquashedMigration
   squashed_migrations *%w[
-    1009015_aggregated_costs_migrations
-    20241120103858_add_start_end_times_to_time_entries
-    20241125104347_add_timezone_identifier_to_time_entry
-    20250219103939_make_time_entry_comment_text_field
-  ].freeze
-
-  tables Tables::CostEntries,
-         Tables::CostTypes,
-         Tables::Rates,
-         Tables::TimeEntries,
-         Tables::TimeEntryActivitiesProjects,
-         Tables::TimeEntryJournals
+    1019015_aggregated_overviews_migrations
+  ]
 end
